LPS-54663 completedCompanies is no bigger than companyIds.length, it should be int, not long.

+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/internal/background/task/ReindexBackgroundTaskStatusMessageTranslator.java
@@ -66,7 +66,7 @@ public class ReindexBackgroundTaskStatusMessageTranslator
 		backgroundTaskStatus.setAttribute(
 			ReindexBackgroundTaskConstants.COUNT, count);
 
-		long completedCompanies = 0;
+		int completedCompanies = 0;
 
 		for (long completedCompanyId : companyIds) {
 			if (completedCompanyId == companyId) {
